refactor: 移动端路由处理器按报表域拆分（mobileRouteHandler.ts 4452 行 → 分发器 156 行）
- handler 实现拆入 src/services/mobileHandlers/（13 个报表域模块 + shared.ts 工具函数），SQL 口径与逻辑零改动
- mobileRouteHandler.ts 仅保留 URL 解析与 case 分发，CI 契约脚本解析格式不变
- check_api_contract.py / check_mobile_routes_coverage.py 同步支持扫描 mobileHandlers/ 目录
- 验证：typecheck / lint(0 errors) / 4 个 CI 门禁脚本 / 生产构建全部通过；40 case、38 handler 与原文件逐一比对一致
